cdktf_azure_sentiment_ml/function/requirements.txt
azure-functions
azure-core
azure-ai-textanalytics
